requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.3.7
python-dotenv==1.0.0
pandas==2.1.0
openpyxl==3.1.2
openai-whisper==20231117
pydub==0.25.1
psycopg2-binary==2.9.9